New setup to support SSR streaming
