+++ src/next-app-router/index.ts
@@ -1,3 +1,3 @@
 export type { RegisterLink } from "../link";
 export type { DefaultColorScheme } from "./zz_internal/defaultColorScheme";
-export { startReactDsfr } from "./zz_internal/start";
+export { DsfrProviderBase, type DsfrProviderProps, StartDsfrOnHydration } from "./DsfrProvider";
