+++ src/next-app-router/server-only-index.ts
@@ -1,0 +1,2 @@
+export { createGetHtmlAttributes } from "./getHtmlAttributes";
+export { DsfrHeadBase, type DsfrHeadProps } from "./DsfrHead";
