+++ src/next-app-router/zz_internal/defaultColorScheme.ts
@@ -1,5 +1,4 @@
 import { assert } from "tsafe/assert";
-
 import type { ColorScheme } from "../../useIsDark";
 
 export type DefaultColorScheme = ColorScheme | "system";
