+++ src/next-app-router/zz_internal/fontUrlByFileBasename.ts
@@ -1,0 +1,23 @@
+import marianneLightWoff2Url from "@codegouvfr/react-dsfr/dsfr/fonts/Marianne-Light.woff2";
+import marianneItalicWoff2Url from "@codegouvfr/react-dsfr/dsfr/fonts/Marianne-Light_Italic.woff2";
+import marianneRegularWoff2Url from "@codegouvfr/react-dsfr/dsfr/fonts/Marianne-Regular.woff2";
+import marianneRegularItalicWoff2Url from "@codegouvfr/react-dsfr/dsfr/fonts/Marianne-Regular_Italic.woff2";
+import marianneMediumWoff2Url from "@codegouvfr/react-dsfr/dsfr/fonts/Marianne-Medium.woff2";
+import marianneMediumItalicWoff2Url from "@codegouvfr/react-dsfr/dsfr/fonts/Marianne-Medium_Italic.woff2";
+import marianneBoldWoff2Url from "@codegouvfr/react-dsfr/dsfr/fonts/Marianne-Bold.woff2";
+import marianneBoldItalicWoff2Url from "@codegouvfr/react-dsfr/dsfr/fonts/Marianne-Bold_Italic.woff2";
+import spectralRegularWoff2Url from "@codegouvfr/react-dsfr/dsfr/fonts/Spectral-Regular.woff2";
+import spectralExtraBoldWoff2Url from "@codegouvfr/react-dsfr/dsfr/fonts/Spectral-ExtraBold.woff2";
+
+export const fontUrlByFileBasename = {
+    "Marianne-Light": marianneLightWoff2Url,
+    "Marianne-Light_Italic": marianneItalicWoff2Url,
+    "Marianne-Regular": marianneRegularWoff2Url,
+    "Marianne-Regular_Italic": marianneRegularItalicWoff2Url,
+    "Marianne-Medium": marianneMediumWoff2Url,
+    "Marianne-Medium_Italic": marianneMediumItalicWoff2Url,
+    "Marianne-Bold": marianneBoldWoff2Url,
+    "Marianne-Bold_Italic": marianneBoldItalicWoff2Url,
+    "Spectral-Regular": spectralRegularWoff2Url,
+    "Spectral-ExtraBold": spectralExtraBoldWoff2Url
+} as const;
